+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Ignore some unnecessary methods defined on some Set implementations in set_spec
 - Run specs on JRuby 10 in Github Actions
 - Retry TruffleRuby rspec runs on Github Actions to avoid random failures due to flakey GC.
+- Extract UNDEFINED to its own file and require it where used.
 
 ## [0.2.0] - 2025-04-09